+++ R/deprecated.R
@@ -3,12 +3,12 @@
 
 
 #' @name defunct
-#' @inherit basejump::defunct
+#' @inherit basejump::defunct description return title
 #' @keywords internal
 NULL
 
 #' @name deprecated
-#' @inherit basejump::deprecated
+#' @inherit basejump::deprecated description return title
 #' @keywords internal
 NULL
 
@@ -19,6 +19,15 @@ NULL
 #' @export
 plotEnrichment <- function(...) {
     .Defunct("plotEnrichedGeneSets")
+}
+
+
+
+## v0.1.15 =====================================================================
+#' @rdname defunct
+#' @export
+plotGSEATable <- function(...) {
+    .Defunct()
 }